dom/manifest/test/test_ManifestProcessor_splash_screens.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1162808
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1162808</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script src="common.js"></script>
  <script>
/**
 * Manifest splash_screens member
 * http://www.w3.org/TR/appmanifest/#splash_screens-member
 **/

'use strict';

typeTests.forEach((type) => {
  var expected = `Expect non-array splash_screens to be empty: ${typeof type}.`;
  data.jsonText = JSON.stringify({
    splash_screens: type
  });
  var result = processor.process(data);
  is(result.splash_screens.length, 0, expected);
});
</script>
</head>
